requirements.txt
accelerate==1.7.0
diffusers==0.33.1
ipykernel==6.29.5
moviepy==2.2.1
openai==1.82.0
openai-whisper==20240930
torchvision==0.21.0+cu118
typing-inspection==0.4.1
wheel==0.45.1
